+++ BTCPayServer.Plugins.LNbank/Pages/Helpers.cs
@@ -6,16 +6,16 @@ namespace BTCPayServer.Plugins.LNbank.Pages;
 
 public static class Helpers
 {
-    public static string Sats(LightMoney amount)
+    public static string Sats(LightMoney amount, bool includeUnit = true)
     {
-        return amount >= LightMoney.Satoshis(1)
-            ? $"{Math.Floor(amount.ToUnit(LightMoneyUnit.Satoshi))} sats"
-            : Millisats(amount);
+        return amount >= LightMoney.Satoshis(1) || amount <= LightMoney.Satoshis(-1)
+            ? $"{Math.Floor(amount.ToUnit(LightMoneyUnit.Satoshi))}{(includeUnit ? " sats" : string.Empty)}"
+            : Millisats(amount, includeUnit);
     }
 
-    public static string Millisats(LightMoney amount)
+    public static string Millisats(LightMoney amount, bool includeUnit = true)
     {
-        return $"{amount.ToUnit(LightMoneyUnit.MilliSatoshi)} millisats";
+        return $"{amount.ToUnit(LightMoneyUnit.MilliSatoshi)}{(includeUnit ? " millisats" : string.Empty)}";
     }
 
     public static string TransactionStateClass(Transaction transaction)