+++ BTCPayServer.Plugins.LNbank/Pages/Wallets/Wallet.cshtml.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using BTCPayServer.Data;
 using BTCPayServer.Plugins.LNbank.Authentication;
 using BTCPayServer.Plugins.LNbank.Data.Models;
@@ -15,12 +14,18 @@ namespace BTCPayServer.Plugins.LNbank.Pages.Wallets;
 [Authorize(AuthenticationSchemes = AuthenticationSchemes.Cookie, Policy = LNbankPolicies.CanViewWallet)]
 public class WalletModel : BasePageModel
 {
+    private readonly HistogramService _histogramService;
     public IEnumerable<Transaction> Transactions { get; set; }
+    public HistogramData HistogramData { get; set; }
 
     public WalletModel(
         UserManager<ApplicationUser> userManager,
+        HistogramService histogramService,
         WalletRepository walletRepository,
-        WalletService walletService) : base(userManager, walletRepository, walletService) { }
+        WalletService walletService) : base(userManager, walletRepository, walletService)
+    {
+        _histogramService = histogramService;
+    }
 
     public IActionResult OnGetAsync(string walletId)
     {
@@ -28,6 +33,7 @@ public class WalletModel : BasePageModel
             return NotFound();
 
         Transactions = CurrentWallet.Transactions.OrderByDescending(t => t.CreatedAt);
+        HistogramData = _histogramService.GetHistogram(CurrentWallet);
 
         return Page();
     }